requirements.txt
fastapi>=0.110,<0.111
uvicorn[standard]>=0.29,<0.30
gunicorn>=21.2,<22

google-api-python-client>=2.129,<3
google-auth>=2.29,<3
google-auth-httplib2>=0.2.0,<0.3.0
google-auth-oauthlib>=1.2.0,<2

redis[hiredis]>=5.0,<6
python-dotenv>=1.0,<2
pydantic>=2.6,<3
httpx>=0.27,<0.28
python-jose[cryptography]>=3.3,<4
itsdangerous>=2.2,<3
tenacity>=8.2,<9
